removed type_as from torch.complex not supported

+++ complex_cnn_model.py
@@ -59,8 +59,7 @@ class ComplexPreActBlock(nn.Module):
 def complex_dropout2d(input, p=0.5, training=True):
     mask = torch.ones(*input.shape, dtype = torch.float32, device=torch.device('cuda'))
     #mask = torch.ones(*input.shape, dtype = torch.float32)
-
-    mask = dropout2d(mask, p, training ) * 1 / (1-p)
+    mask = dropout2d(mask, p, training) * 1 / (1-p)
     mask.type(input.dtype)
     return mask * input
 
@@ -131,7 +130,7 @@ class ComplexPreActResNetFFTKnee(nn.Module):
     def forward(self, kspace):
         print("the kspace shape is {}".format(kspace.shape)) # torch.size([8, 1, 640, 400])
         if self.data_space == 'complex_input':
-            out = torch.complex(kspace.real, kspace.imag).type(torch.complex64).type_as(kspace)
+            out = torch.complex(kspace.real, kspace.imag).type(torch.complex64)
             #out = torch.complex(kspace.real, kspace.imag).type(torch.complex64)
             out = center_crop(out, self.image_shape)
             out = self.conv_comp(out)
